% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqts-acquisition.R
\name{append}
\alias{append}
\alias{aqAppendTimeseries}
\alias{aqOverwriteAppendTimeseries}
\alias{aqAppendReflectedTimeseries}
\title{Append Timeseries}
\usage{
aqAppendTimeseries(UniqueId, Points, .perform = TRUE)

aqOverwriteAppendTimeseries(
  UniqueId,
  Start,
  End,
  Points,
  open = TRUE,
  .perform = TRUE
)

aqAppendReflectedTimeseries(
  UniqueId,
  Start,
  End,
  Points,
  open = TRUE,
  .perform = TRUE
)
}
\arguments{
\item{UniqueId}{The guid of the timeseries}

\item{Points}{a data.frame with data to write; must have at least \code{Time} and \code{Value} columns; \code{Type} ("Point" or "Gap"), \code{GradeCode} (numeric), and \code{Qualifiers} (character vector) are also allowed}

\item{.perform}{if TRUE, perform the query and return the response}

\item{Start}{the beginning of the time range to overwrite}

\item{End}{the end of the time range to overwrite}

\item{open}{should the \code{End} argument be considered an open (exclusive) or closed (inclusive) interval}
}
\description{
These functions append data to timeseries. \code{aqAppendTimeSeries} can be used
to append data to basic timeseries. No existing data will be overwritten.
\code{aqOverwriteAppendTimeSeries} will overwrite existing data
}
