% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquarius-ts-token.R
\name{aq_token}
\alias{aq_token}
\title{Create an Aquarius web service token}
\usage{
aq_token(access_token, expires_in = 3600, .date = Sys.time())
}
\arguments{
\item{access_token}{The access token used to authenticate request}

\item{expires_in}{Number of seconds until token expires.}

\item{.date}{Date the request was made; used to convert the relative
\code{expires_in} to an absolute \code{expires_at}.}
}
\value{
An X-Authentication-Token for Aquarius Time Series: an S3 list with class \code{aq_token}.
}
\description{
Creates a S3 object of class \verb{<aq_token>} representing a token
returned from POST /publish/v2/session.
}
\examples{
aq_token("abcdef")
aq_token("abcdef", expires_in = 3600)
}
