% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{ExtendedFilters}
\alias{ExtendedFilters}
\title{Formats the ExtendedFilters argument of GetLocationDescriptionList and GetTimeSeriesDescriptionList
The ExtendedFilters argument uses the ServiceStack jsv encoding}
\usage{
ExtendedFilters(FilterName, FilterValue)
}
\description{
Formats the ExtendedFilters argument of GetLocationDescriptionList and GetTimeSeriesDescriptionList
The ExtendedFilters argument uses the ServiceStack jsv encoding
}
