% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquarius-ts-publish.R
\name{fieldvisit-requests}
\alias{fieldvisit-requests}
\alias{GetFieldVisitDescriptionList}
\alias{GetFieldVisitData}
\alias{GetFieldVisitDataByLocation}
\alias{GetFieldVisitReadingsByLocation}
\title{Publish API: Functions for Field Visits}
\usage{
GetFieldVisitDescriptionList(
  ...,
  LocationIdentifier = NULL,
  QueryFrom = NULL,
  QueryTo = NULL,
  .format = TRUE,
  .perform = TRUE
)

GetFieldVisitData(FieldVisitIdentifier, ..., .format = TRUE, .perform = TRUE)

GetFieldVisitDataByLocation(
  LocationIdentifier,
  Parameters = NULL,
  Activities = NULL,
  ...,
  .format = TRUE,
  .perform = TRUE
)

GetFieldVisitReadingsByLocation(
  LocationIdentifier,
  LocationUniqueId,
  Parameters,
  ...,
  .format = TRUE,
  .perform = TRUE
)
}
\arguments{
\item{...}{pass query arguments to route. Please see Publish API documentation
for available arguments.}
}
\description{
These functions retrieve field visit information via the AQTS Publish API
}
