% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webportal.R
\name{GetExport}
\alias{GetExport}
\title{Read from the Web Portal export tab}
\usage{
GetExport(...)
}
\description{
This function reads from the Web Portal export tab.
https://aquarius.portlandoregon.gov/Export/BulkExport?DateRange=EntirePeriodOfRecord&TimeZone=-8&Calendar=CALENDARYEAR&Interval=PointsAsRecorded&Step=1&ExportFormat=csv&TimeAligned=False&RoundData=False&IncludeGradeCodes=True&IncludeApprovalLevels=True&IncludeQualifiers=True&IncludeInterpolationTypes=True&Datasets[0].DatasetName=Temperature.7DADM%40P0337&Datasets[0].Calculation=Instantaneous&Datasets[0].UnitId=168&_=1686947628225
}
