% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqts-provisioning.R
\name{getLocationDatum}
\alias{getLocationDatum}
\title{Get items nested under locations}
\usage{
getLocationDatum(LocationUniqueId, query = "/Results", .perform = TRUE)
}
\arguments{
\item{LocationUniqueId}{the guid for a location}
}
\description{
Get items nested under locations
}
