% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqts-provisioning.R
\name{getLocationTimeseries}
\alias{getLocationTimeseries}
\title{Get items nested under locations}
\usage{
getLocationTimeseries(LocationUniqueId, .perform = TRUE)
}
\arguments{
\item{LocationUniqueId}{the guid for a location}
}
\description{
Get items nested under locations
}
