% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquarius-ts-publish.R
\name{GetReportList}
\alias{GetReportList}
\title{Query Reports}
\usage{
GetReportList(
  LocationUniqueId = NULL,
  TimeSeriesUniqueIds = NULL,
  ReportTitle = NULL,
  ...,
  .format = TRUE,
  .perform = TRUE
)
}
\description{
Queries a
}
