% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish.R
\name{location-requests}
\alias{location-requests}
\alias{GetLocationDescriptionList}
\alias{GetLocationData}
\title{Functions for Locations}
\usage{
GetLocationDescriptionList(
  LocationIdentifier,
  ...,
  .format = TRUE,
  .perform = TRUE
)

GetLocationData(LocationIdentifier, ..., .format = TRUE, .perform = TRUE)
}
\arguments{
\item{...}{pass query arguments to route. Please see Publish API documentation
for available arguments.}
}
\description{
These functions retrieve location information.
}
