% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqts-publish.R
\name{perform_and_format}
\alias{perform_and_format}
\title{Shared parameter definitions}
\arguments{
\item{...}{additional arguments passed to the endpoint - please see the AQTS API documentation for available arguments}

\item{.perform}{if TRUE, perform the query and return the response}

\item{.format}{if TRUE and \code{.perform = TRUE}, format the response, usually as a \code{data.frame}}
}
\description{
Shared parameter definitions
}
