% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqts-acquisition.R
\name{post-attachments}
\alias{post-attachments}
\alias{aqPostReportAttachment}
\alias{aqDeleteReportAttachment}
\alias{aqPostLocationAttachment}
\title{Post Attachements}
\usage{
aqPostReportAttachment(
  File,
  LocationUniqueId,
  Title,
  type,
  ...,
  .perform = TRUE
)

aqDeleteReportAttachment(ReportUniqueId, .perform = TRUE)

aqPostLocationAttachment(File, LocationUniqueId, ..., .perform = TRUE)
}
\arguments{
\item{...}{pass query arguments to route. Please see Acquisition API documentation
for available arguments.}
}
\description{
These functions post files as attachments.
}
