% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquarius-ts-provisioning.R
\name{prDeleteLocation}
\alias{prDeleteLocation}
\title{Delete a location}
\usage{
prDeleteLocation(unique_id, .perform = TRUE)
}
\description{
Deletes a monitoring location. Only an empty location can be deleted.
}
