% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquarius-ts-publish.R
\docType{data}
\name{publish-config}
\alias{publish-config}
\alias{GetParameterList}
\alias{GetMonitoringMethodList}
\alias{GetUnitList}
\alias{GetApprovalList}
\alias{GetGradeList}
\alias{GetQualifierList}
\alias{TagApplicability}
\alias{GetTagList}
\alias{GetActiveMetersAndCalibrations}
\title{Functions for retrieving configuration items}
\format{
An object of class \code{character} of length 5.
}
\usage{
GetParameterList(.format = TRUE, .perform = TRUE)

GetMonitoringMethodList(.format = TRUE, .perform = TRUE)

GetUnitList(GroupIdentifier = NULL, .format = TRUE, .perform = TRUE)

GetApprovalList(.format = TRUE, .perform = TRUE)

GetGradeList(.format = TRUE, .perform = TRUE)

GetQualifierList(.format = TRUE, .perform = TRUE)

TagApplicability

GetTagList(Applicability = NULL, .format = TRUE, .perform = TRUE)

GetActiveMetersAndCalibrations(.format = TRUE, .perform = TRUE)
}
\arguments{
\item{GroupIdentifier}{Filter to a specific Unit Group.}

\item{Applicability}{tag applicability, one of  "AppliesToLocations",
"AppliesToLocationNotes", "AppliesToSensorsGauges","AppliesToAttachements",
"AppliesToReports"}

\item{...}{pass query arguments to route. Please see Publish API documentation
for available arguments.}
}
\description{
These functions retrieve "system-wide configuration items". These are look up
lists that are used as allowed values throughout Aquarius.
}
\section{Functions}{
\itemize{
\item \code{GetParameterList()}: Retrieve parameter list.

\item \code{GetMonitoringMethodList()}: Retrieve monitoring methods.

\item \code{GetUnitList()}: Retrieve units.

\item \code{GetApprovalList()}: Retrieve approvals.

\item \code{GetGradeList()}: Retrieve grades.

\item \code{GetQualifierList()}: Retrieve qualifiers list.

\item \code{TagApplicability}: tag applicability

\item \code{GetTagList()}: Retrieve tags

}}
\keyword{datasets}
