% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raquarius-package.R
\docType{package}
\name{raquarius-package}
\alias{raquarius}
\alias{raquarius-package}
\title{raquarius: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jasonelaw/raquarius}
  \item Report bugs at \url{https://github.com/jasonelaw/raquarius/issues}
}

}
\keyword{internal}
