% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish.R
\name{ts-requests}
\alias{ts-requests}
\alias{GetTimeSeriesUniqueIdList}
\alias{GetTimeSeriesDescriptionListByUniqueId}
\alias{GetTimeSeriesData}
\alias{GetApprovalsTransactionList}
\alias{GetTimeSeriesCorrectedData}
\alias{GetMetadataChangeTransactionList}
\title{Retrieve Time Series Information From Publish API}
\usage{
GetTimeSeriesUniqueIdList(
  LocationIdentifier,
  Parameter,
  ...,
  .format = TRUE,
  .perform = TRUE
)

GetTimeSeriesDescriptionListByUniqueId(
  TimeSeriesUniqueIds,
  .format = TRUE,
  .perform = TRUE
)

GetTimeSeriesData(
  TimeSeriesUniqueIds,
  QueryFrom,
  QueryTo,
  ...,
  .format = TRUE,
  .perform = TRUE
)

GetApprovalsTransactionList(
  TimeSeriesUniqueId,
  QueryFrom,
  QueryTo,
  ...,
  .format = TRUE,
  .perform = TRUE
)

GetTimeSeriesCorrectedData(
  TimeSeriesUniqueId,
  QueryFrom,
  QueryTo,
  Unit,
  ...,
  .format = TRUE,
  .perform = TRUE
)

GetMetadataChangeTransactionList(
  TimeSeriesUniqueId,
  ...,
  .perform = TRUE,
  .format = TRUE
)
}
\arguments{
\item{...}{pass query arguments to route. Please see Publish API documentation
for available arguments.}

\item{TimeSeriesUniqueId}{a unique id}
}
\description{
These functions retrieve time series information.
}
