% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webportal.R
\name{webportal-routes}
\alias{webportal-routes}
\alias{GetMapDataAllLocations}
\alias{GetMapDataDatasetsByParameter}
\alias{GetExportDataSet}
\alias{GetExportTimeAligned}
\title{Web Portal Routes}
\usage{
GetMapDataAllLocations(...)

GetMapDataDatasetsByParameter(parameter, ...)

GetExportDataSet(...)

GetExportTimeAligned(
  Datasets,
  Calendar = NULL,
  StartTime = NULL,
  EndTime = NULL,
  Step = NULL,
  Timezone = NULL,
  DateRange = c("EntirePeriodOfRecord", "OverlappingPeriodOfRecord", "Today", "Days7",
    "Days30", "Months6", "Years1"),
  Interval = c("PointsAsRecorded", "Minutely", "Hourly", "Daily", "Monthly", "Yearly"),
  RoundData = FALSE,
  IncludeGradeCodes = TRUE,
  IncludeQualifiers = TRUE,
  IncludeApprovalLevels = TRUE,
  IncludeInterpolationTypes = TRUE,
  ...
)
}
\description{
See the web portal API documentation for arguments.
#' @param ... pass query arguments to route. Please see Web Portal API documentation
for available arguments.
}
\section{Functions}{
\itemize{
\item \code{GetMapDataAllLocations()}: List all locations

\item \code{GetMapDataDatasetsByParameter()}: List timeseries by parameter

\item \code{GetExportDataSet()}: Export a data set

\item \code{GetExportTimeAligned()}: Export several time-aligned data sets.

}}
