% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquarius-ts-acquisition.R
\name{post-attachments}
\alias{post-attachments}
\alias{PostReportAttachment}
\alias{DeleteReportAttachment}
\alias{PostLocationAttachment}
\title{Post Attachements}
\usage{
PostReportAttachment(File, LocationUniqueId, Title, type, ..., .perform = TRUE)

DeleteReportAttachment(ReportUniqueId, .perform = TRUE)

PostLocationAttachment(File, LocationUniqueId, ..., .perform = TRUE)
}
\arguments{
\item{...}{pass query arguments to route. Please see Acquisition API documentation
for available arguments.}
}
\description{
These functions post files as attachments.
}
