% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish.R
\name{publish-config}
\alias{publish-config}
\alias{GetParameterList}
\alias{GetMonitoringMethodList}
\alias{GetUnitList}
\alias{GetApprovalList}
\alias{GetGradeList}
\alias{GetQualifierList}
\alias{GetTagList}
\alias{GetActiveMetersAndCalibrations}
\title{Functions for retrieving configuration items}
\usage{
GetParameterList()

GetMonitoringMethodList()

GetUnitList(...)

GetApprovalList()

GetGradeList()

GetQualifierList()

GetTagList(...)

GetActiveMetersAndCalibrations()
}
\arguments{
\item{...}{pass query arguments to route. Please see Publish API documentation
for available arguments.}
}
\description{
These functions retrieve "system-wide configuration items". These are look up
lists that are used as allowed values throughout Aquarius.
}
\section{Functions}{
\itemize{
\item \code{GetParameterList()}: Retrieve parameter list.

\item \code{GetMonitoringMethodList()}: Retrieve monitoring methods.

\item \code{GetUnitList()}: Retrieve units.

\item \code{GetApprovalList()}: Retrieve approvals.

\item \code{GetGradeList()}: Retrieve grades.

\item \code{GetQualifierList()}: Retrieve qualifiers list.

\item \code{GetTagList()}: Retrieve tags.

}}
