% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raquarius-package.R
\docType{package}
\name{raquarius-package}
\alias{raquarius}
\alias{raquarius-package}
\title{raquarius: Aquarius Time Series API Client For R}
\description{
An interface to the API for Aquarius Time Series. The Publish API is currently mostly supported. In the future, it may support the Provisioning and Acquisition API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jasonelaw/raquarius}
  \item Report bugs at \url{https://github.com/jasonelaw/raquarius/issues}
}

}
\keyword{internal}
