% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish.R
\name{ts-requests}
\alias{ts-requests}
\alias{GetTimeSeriesUniqueIdList}
\alias{GetTimeSeriesDescriptionListByUniqueId}
\alias{GetTimeSeriesData}
\alias{GetApprovalsTransactionList}
\alias{GetTimeSeriesCorrectedData}
\title{Retrieve Locations From Publish API}
\usage{
GetTimeSeriesUniqueIdList(...)

GetTimeSeriesDescriptionListByUniqueId(req, TimeSeriesUniqueIds)

GetTimeSeriesData(...)

GetApprovalsTransactionList(...)

GetTimeSeriesCorrectedData(...)
}
\arguments{
\item{...}{pass query arguments to route. Please see Publish API documentation
for available arguments.}
}
\description{
These functions retrieve location information.
}
