% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquarius-webportal.R
\name{webportal-routes}
\alias{webportal-routes}
\alias{GetFilters}
\alias{GetLocation}
\alias{GetLocations}
\alias{GetDataset}
\alias{GetLatestStatistics}
\alias{GetLatestStatistic}
\alias{GetLatestStatisticsByParameter}
\alias{GetLatestStatisticValues}
\alias{GetLatestStatisticValuesByParameter}
\alias{GetLatestStatisticValuesByStatistic}
\alias{GetMapDataAllLocations}
\alias{GetMapDataDatasetsByParameter}
\alias{GetMapDataLatestStatistics}
\alias{GetMapDataPeriodicStatistics}
\alias{GetExportDataSet}
\alias{GetExportSeasonalStatistic}
\alias{GetExportTimeAligned}
\alias{GetExportBulk}
\title{Web Portal API}
\usage{
GetFilters(.perform = TRUE, .format = TRUE)

GetLocation(location, ..., .format = TRUE, .perform = TRUE)

GetLocations(..., .format = TRUE, .perform = TRUE)

GetDataset(..., .perform = TRUE, .format = TRUE)

GetLatestStatistics(
  parameter = NULL,
  statistic = NULL,
  active = NULL,
  .format = TRUE,
  .perform = TRUE
)

GetLatestStatistic(
  parameter = NULL,
  statistic = NULL,
  ...,
  .format = TRUE,
  .perform = TRUE
)

GetLatestStatisticsByParameter(
  parameter = NULL,
  ...,
  .format = TRUE,
  .perform = TRUE
)

GetLatestStatisticValues(..., .format = TRUE, .perform = TRUE)

GetLatestStatisticValuesByParameter(
  parameter,
  ...,
  .format = TRUE,
  .perform = TRUE
)

GetLatestStatisticValuesByStatistic(
  parameter,
  statistic,
  ...,
  .format = TRUE,
  .perform = TRUE
)

GetMapDataAllLocations(..., .perform = TRUE, .format = TRUE)

GetMapDataDatasetsByParameter(parameter, ..., .perform = TRUE, .format = TRUE)

GetMapDataLatestStatistics(
  parameter,
  statistic,
  ...,
  .perform = TRUE,
  .format = TRUE
)

GetMapDataPeriodicStatistics(
  parameter,
  statistic,
  interval,
  date,
  ...,
  .perform = TRUE,
  .format = TRUE
)

GetExportDataSet(DataSet, ..., .perform = TRUE, .format = TRUE)

GetExportSeasonalStatistic(
  dataset = NULL,
  interval = NULL,
  statistic = NULL,
  referenceperiod = NULL,
  ...,
  .perform = TRUE,
  .format = TRUE
)

GetExportTimeAligned(
  Datasets,
  Calendar = NULL,
  StartTime = NULL,
  EndTime = NULL,
  Step = NULL,
  Timezone = NULL,
  DateRange = c("EntirePeriodOfRecord", "OverlappingPeriodOfRecord", "Today", "Days7",
    "Days30", "Months6", "Years1"),
  Interval = c("PointsAsRecorded", "Minutely", "Hourly", "Daily", "Monthly", "Yearly"),
  RoundData = FALSE,
  IncludeGradeCodes = TRUE,
  IncludeQualifiers = TRUE,
  IncludeApprovalLevels = TRUE,
  IncludeInterpolationTypes = TRUE,
  ...
)

GetExportBulk(..., .perform = TRUE, .format = TRUE)
}
\arguments{
\item{.perform}{if \code{TRUE}, perform the request, otherwise return a
\code{httr2::request} object}

\item{.format}{if \code{TRUE}, return a \code{data.frame}, otherwise return a
\code{httr2::response} object. Ignored if \code{.perform = FALSE}}

\item{...}{pass query arguments to route. Please see Web Portal API
documenation for available arguments.}
}
\description{
See the web portal API documentation for arguments.
}
\section{Functions}{
\itemize{
\item \code{GetFilters()}: Get filters for map endpoints

\item \code{GetLocation()}: Get a location

\item \code{GetLocations()}: Get locations

\item \code{GetDataset()}: Get datasets

\item \code{GetLatestStatistics()}: Get Latest Statistics

\item \code{GetLatestStatistic()}: Get Latest Statistics

\item \code{GetLatestStatisticsByParameter()}: Get Latest Statistics

\item \code{GetLatestStatisticValues()}: Get Latest Statistics

\item \code{GetLatestStatisticValuesByParameter()}: Get Latest Statistics

\item \code{GetLatestStatisticValuesByStatistic()}: Get Latest Statistics

\item \code{GetMapDataAllLocations()}: List all locations

\item \code{GetMapDataDatasetsByParameter()}: List timeseries by parameter; return sf object

\item \code{GetMapDataLatestStatistics()}: Get latest statistics; return sf object

\item \code{GetMapDataPeriodicStatistics()}: Get periodic statistics; return sf object

\item \code{GetExportDataSet()}: Export a data set

\item \code{GetExportSeasonalStatistic()}: Export seasonal statistics

\item \code{GetExportTimeAligned()}: Export several time-aligned data sets.

\item \code{GetExportBulk()}: Export datasets in bulk; optionally summarize

}}
